split js into multiple files (socket io error)
