+++ public/static/js/session/save_latex.js
@@ -1,0 +1,6 @@
+$(function() {
+	$("#saveLatex").click(function(e) {
+		e.preventDefault();
+		socket.emit('latexSaved', {'rawText': $("#latexInput").val(), 'sendingUser': window.name});
+	});
+});
